project cleanup

+++ src/common/data/data_model/data_model.cpp
@@ -97,11 +97,11 @@ const char *DataModel::model_type_extended() const {
 std::string DataModel::frame_accessor(const char * field) {
   // if the current model is an array we will return an index
   if (_type == DATA_MODEL_CODE_ARRAY) {
-    std::string r_val(sizeof(size_t), '\0');
+    std::string r_val(sizeof(long), '\0');
 
-    for (size_t i=0; i<_fields->size(); i++) {
+    for (long i=0; i<_fields->size(); i++) {
       if (strcmp(_fields->at(i)->name(), field) == 0) {
-        *((size_t*)r_val.c_str()) = i;
+        *((long*)r_val.c_str()) = i;
         return r_val;
       }
     }
